Allow testing of operations, make operations public
For now we'll leave them public. It's the simplest solution, really. Testing private functions is so obtuse with Rust's tooling at the moment that I'd rather just make them public and document that these functions are for internal use, though it's still safe to call them really.
This makes doctests and unit tests and integration tests so much easier.