Move all tests into the source directory
This goes against the common Unix project structure a little bit. The source code should contain just the source code, not the tests. However this does follow Rust's code structure guidelines which says that the unit tests should be placed inside of the files that contain the symbols they are testing.
This is done now in order to be able to access the private members of structs that we are testing. In order to access the vertices vector of a polygon, we'll need to be inside of the polygon.rs file or in an adjacent subdirectory. It's a little sad that this has to be done this way, but understandable.
The integration tests will now be placed inside of the source code as well. They can't access the private members from a tests submodule but at least they'll be okay there.
I'll look into making private detail functions private again in the future. Maybe that is now possible again with this structure.
Helper functions (such as the polygon fixtures) will be placed in the tests module too so that we can access them via crate::tests::data::polygon. In the tests directory next to the source code they were inaccessible from within the source code directory.