Implement sincing back from GPU to host
This would complete the loop.
Only these functions are still untested. They are untested because they wouldn't compile without the interior mutability pattern. Functions that shouldn't modify the polygon (such as len()) would modify the polygon by triggering a sync from the GPU to the host. This can't be mutating it since it wouldn't implement the traits any more.